+++ src/deepkeel/contrib/postgres/__init__.py
@@ -1,0 +1,31 @@
+"""Installable PostgreSQL adapters for DeepKeel production ports."""
+
+from deepkeel.contrib.postgres.checkpoint_store import PostgresDurableCheckpointStore
+from deepkeel.contrib.postgres.bundle import PostgresRuntimeBundle
+from deepkeel.contrib.postgres.database import PostgresDatabase
+from deepkeel.contrib.postgres.event_journal import PostgresRuntimeEventJournal
+from deepkeel.contrib.postgres.governance import (
+    PostgresBudgetLedger,
+    PostgresModelHealthStore,
+    PostgresRunControl,
+)
+from deepkeel.contrib.postgres.lease_store import PostgresRunLeaseStore
+from deepkeel.contrib.postgres.model_store import PostgresModelInvocationStore
+from deepkeel.contrib.postgres.state_store import PostgresRuntimeStateStore
+from deepkeel.contrib.postgres.tool_store import PostgresToolExecutionStore
+from deepkeel.contrib.postgres.trace_store import PostgresTraceStore
+
+__all__ = [
+    "PostgresDatabase",
+    "PostgresBudgetLedger",
+    "PostgresDurableCheckpointStore",
+    "PostgresModelHealthStore",
+    "PostgresModelInvocationStore",
+    "PostgresRunLeaseStore",
+    "PostgresRunControl",
+    "PostgresRuntimeBundle",
+    "PostgresRuntimeEventJournal",
+    "PostgresRuntimeStateStore",
+    "PostgresToolExecutionStore",
+    "PostgresTraceStore",
+]